refactor for shorter components
